Replace tools/if_not_there.c with shell script.
When building Mercury natively with MinGW-w64 x86_64 under MSYS2
I saw that some process would crash in the boehm_gc directory:

  ./if_not_there mach_dep.o /bin/sh [snip]/../scripts/mgnuc [snip] ./mach_dep.c
  ^^^^Starting command^^^^

  This application has requested the Runtime to terminate it in an unusual way.
  Please contact the application's support team for more information.

The same command would not crash if I ran it manually in the shell,
though it did return to the shell before completing.

Replacing if_not_there.c by a shell script avoids the problem, such that
I was able to continue and build the rest of the compiler.  We require a
working shell on all platforms that we support, unlike Boehm GC upstream,
so I'm going to leave it at that.

tools/if_not_there.sh:
	Add shell script that mimics if_not_there.c.

Makefile.direct:
	Use if_not_there.sh in place of the compiled version of
	if_not_there.c.