Add GC_mercury_write_size_map()
Add a function to print the size mapping information, that is for each
request size the actual size allocated (both in words).

This function was created with the original code for enumerating the
reachable heap objects, however when that code was ported upstream this was
not.  Add it now for compatibility with Mercury's heap profiling grade.

This code was probably originally written by Peter Wang, but I'm not sure.

include/gc.h:
misc.c:
    As above.